plaid: A playlist and dmenu generator all in one
Introducing `plaid`.  This little tool is designed to solve the problem
of downloading multitudes of playlists which consume large amounts of
harddrive space unnecessarily.

It works by taking a known url, such as youtube.com and generates a
playlist for `mpv`.  It also adds an entry in a file that can be loaded
with `dmenu_loader` to produce a dmenu of all the current playlists with
pretty titles.

Because I am generating real playlist files, mpv can be exploited to
save and restore state thus keeping track of where we were mid-stream
and resuming playlists where we left off.