ditch: Support loading stream/team data from files
This reworks the structure to support XDG_CONFIG_HOME/ditch with the
following contents:

ditch/streamers
ditch/teams

Where each file contains a list of respective elements.
It uses a chunking method for dealing with large amounts of streamers
otherwise both curl, primarily, and twitch will either fail or not
accept it.

Another small change is how I distinguish between team and stream data.
My current code always appends '&limit=' to stream data which allows me
to depend on this existing.  It's not ideal but it'll do.