dircolors: Actually define TERM linux
Because TERM linux wasn't available on login, dircolors would choke on
the unknown sequences on the tty (TERM=linux) and subsequently set
LS_COLORS to nothing.  This meant that `ls --color=auto` worked on the
tty because the /default/ dircolors recognises TERM=linux but would fail
when using a TE on Xorg.  Why you do this coreutils 8.23? ._.