ditch: Clean ups and logic improvements.
Instead of using regex to capture segments of the string, simply use the
builtin functionalty of read and positional arguments instead, this
reduces a lot of the syntax-bloat needed and probably makes it more
effecient (not that it matters with bash).

Fixed up a few comments, reorded and changed some logic and tidied up a
little.