repo-cycle-detector: Improve cycle results.
Added a small helper function to constrain the results of expac -S to
just the specified repo.

I also print out the tsorted order as it might be useful.  Errors print
on stderr so adding > /dev/null to the innvocation should be suitable if
output isn't wanted but only errors.