user-dirs.dir: specifically set xdg_desktop_dir
The xdg-user-dirs tool will literally check if the 6 bytes after a '='
is equal to "$HOME" to determine if it is relative or not.  If "$HOME"
is not present it assumes all paths are relative to the users home.  If
any of these XDG_*_DIR keys are not set it will default to the users
home except XDG_DESKTOP_DIR for "historical reasons":

    /* Special case desktop for historical compatibility */
    if (strcmp (type, "DESKTOP") == 0)
      {
        user_dir = (char*) malloc (strlen (home_dir) + strlen ("/Desktop") + 1);
        if (user_dir == NULL)
          return NULL;

        strcpy (user_dir, home_dir);
        strcat (user_dir, "/Desktop");
        return user_dir;
      }

So I need to specifically set the XDG_DESKTOP_DIR to "$HOME" in order to
prevent firefox from creating a ~/Desktop directory.