ecks: configure xauth prior to starting the server
It turns out that if neither xauth or xhost are configured prior to
starting the xserver then it will by default accept all connections even
if xhost is used to include the localuser after the fact.

Because of this we need to configure xauth prior to starting the server
guaranteeing that the server will only accept connections from the host
and mcookie defined in the xauth list.

I've opted to only supporting a single xauthority file for both client
and server for simplicity until there is a need for separation.

Thanks to irc://freenode!grawity for explaining this behaviour.