ditch: switch to tab separated output with jq
This changes removes in the in-band processing of characters that could
appear in the source material.  Literal tabs in jq are converted to a
sequence of spaces allowing tools like sort and column to safely operate
on a literal tab for their delimiters.

This is one step closer to providing a pure command-line approach using
shell completion systems instead of depending on any external menu.