ditch: transition to v5 API
Oh what a mess.  The new API has added UIDs (which are used for channel
IDs as well) which are now needed to query live streamers and such.
This means they now have an ID, a username and a display name.
Strangely the UIDs are returned as *strings* in the json and at this
point I have absolutely no bloody clue what the point of all this was.

Nevertheless, due to this and my distaste for caching, I've added a new
tool called (unhelpfully) n2i. It converts twitch usernames from stdin
to UIDs which I can redirect to my de facto streamers file I use for
ditch.

[I wish I could rationalise why UIDs were introduced for my own sanity.]