ecks: fix race condition between trap handlers
There was a time window between Xorg starting and us reinstating the
SIGUSR1 handler where SIGUSR1 signals would be ignored.

Instead we use a subshell and set the SIG_IGN disposition inside while
also exec'ing Xorg so we can get the correct pid out the end.

Thanks to @grawity for noticing this issue.