asr: handle single satisfies cases
If we get only a single package as a potential provider then we can try
and be a little more helpful by telling the user about the situation and
checking out the package.

This also cleaned up some logic in dealing with nonextant packages by
letting it fall through to the satisfier check.