switching to xterm
My xterm includes a modified xterm-needle.patch[0] based on work by
pbrisbin[1] and fmang[2] which is used to allow keyboard driven url
selection (although it is more general than that).  This improves upon
termite as it allows multiple urls to be matched even if they appear on
the same row.

Additionally xterm has lower input latency[3] and seemingly supports
truecolour (although I can't find an exact patch and those which mention
it refer to aproximation).  However the test scripts from XVilka's[4]
gist appear to work (at least when compared to vte terminals).

I suspect trickery as the official format produces buggy results but the
common semi-colon format works.  It is a lot slower than vte and seems
to redraw whenever switching workspaces to and from the terminal.

Alternatively to xterm, st would be nice but even with tmux there isn't
yet a nice way to do keyboard driven url selection (tmux-url-select is
a bit buggy for me).  Perhaps tmux's copy-mode can be scripted to work in
much the same way as the needle patch.

Until something better comes along, it seems like xterm is the terminal
emulator for me.

0. https://github.com/Earnestly/pkgbuilds/blob/master/xterm-git/xterm-needle.patch
1. https://pbrisbin.com/posts/selecting_urls_via_keyboard_in_xterm
2. https://bbs.archlinux.org/viewtopic.php?id=181515
3. https://lwn.net/Articles/751763/
4. https://gist.github.com/XVilka/8346728