qmatrix: spit and polish
qmatrix has been (or is) a bit of a mess, a quick bodge to implement
rudimentary umatrix-like blocking in qutebrowser.

This change introduces a degree of polish, some clearer thinking about
rule matching and performance improvements.

The '=' flag is gone, this means qmatrix-rules will have to specify both
domains everytime. It was always an awkward duck and needed special
attention everywhere it was used in the code.

Introduction of the '*' flag which enables all resource types.

Switching to compiled regex while still exposing the interface as
fnmatch globs by using fnmatch.translate and then compiling them as
regex pattern objects directly in the ruleset. This change roughly
halved the time it took matching each per-rule (only 60 rules).

Finally, use of SimpleNamespace to keep all of my global state under the
qmatrix namespace.